fixed RPM calculation in move() - minor changes to main, logger and monitor

+++ include/Logger.h
@@ -18,7 +18,7 @@ public:
     ~Logger();
     
     void printXYAccelMag(float val);
-    void printAction(int val);
+    void printAction(float val);
     void printNewDesiredLinearVel(float val);
     void printPitch(float val);
